Add files via upload
SBOT the rescue robot (hexapod) is capable of traveling across uneven terrains using its highly adjustable legs. The primary objective is to deploy this robot in disaster situations, such as locating survivors trapped under collapsed buildings. Our robot is equipped with a range of sensors to aid in finding survivors, including an MQ135 sensor for CO2 monitoring, and features a live camera feed with both manual and autonomous control modes.In addition to its rescue functions, the robot can also serve as a safety device in households located in disaster-prone areas. It is equipped with sensors to detect earthquakes and uses a web server for interfacing with the robot. The robot operates with a communication method that does not rely on internet services, ensuring reliable performance even during disaster scenarios when the internet may be unavailable